feature: support kv binding in xbot
Previously, all configurations are put inside worker secrets, which are
invisible once they're set. This is not a good workflow because they're
not secrets and I simply don't want to put it inside this open repo.

This commit integrates the cloudflare kv storage so I can better manage
the configuration. Integrate kv also makes the serverless bot possible
to maintain some states.