feature(image): upload images to cloudflare r2 (#23)
This commit uploads the received images from telegram bot to cloudflare
r2 storage.

This is helpful because when I made a screenshot, I can send a message
to telegram bot and it will help me put the images into r2. I don't
need to save the screentshot somewhere and then upload it via web ui.