feature(poe): support basic poe echo bot
This commit adds an echo poe bot, which allows to interact with xbot server
via poe.com website. It enables the future more complex RAG workflow.

It also adds a new router to route the requests to corresponding handler.
It serves '/' and '/telegram' for telegram message, and '/poe' for poe bot.