feature(poe): authentication check and add gemini support
This commit adds authentication to ensure only requests come from poe platform
will be served; the permession of the poe bot will be managed in platform side.

It also adds the gemini-2.0-flash as the default model for the poe bot, now it
could serve basic queries.

The context maybe not be maintained well now, and will be done in the future.